Restrict recipe access to authenticated users, and update/delete methods to recipe's author.

+++ src/main/java/io/mayconfrr/recipes/recipe/RecipeController.java
@@ -1,10 +1,14 @@
 package io.mayconfrr.recipes.recipe;
 
+import io.mayconfrr.recipes.user.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.Map;
@@ -13,6 +17,7 @@ import java.util.Map;
 @RequestMapping("/api/recipe")
 @RequiredArgsConstructor
 @Validated
+@PreAuthorize("isAuthenticated()")
 public class RecipeController {
     private final RecipeService recipeService;
 
@@ -22,19 +27,27 @@ public class RecipeController {
     }
 
     @PostMapping("/new")
-    public Map<String, Long> postRecipe(@Valid @RequestBody Recipe recipe) {
+    public Map<String, Long> postRecipe(@Valid @RequestBody Recipe recipe, @AuthenticationPrincipal User user) {
+        recipe.setAuthor(user);
         return Map.of("id", recipeService.saveRecipe(recipe));
     }
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteRecipe(@PathVariable Long id) {
+    public void deleteRecipe(@PathVariable Long id, @AuthenticationPrincipal User user) {
+        if (!recipeService.getRecipeById(id).getAuthor().equals(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         recipeService.deleteRecipeById(id);
     }
 
     @PutMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void updateRecipe(@PathVariable Long id, @Valid @RequestBody Recipe recipe) {
+    public void updateRecipe(@PathVariable Long id, @Valid @RequestBody Recipe recipe, @AuthenticationPrincipal User user) {
+        if (!recipeService.getRecipeById(id).getAuthor().equals(user)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+        recipe.setAuthor(user);
         recipeService.updateRecipeById(id, recipe);
     }
